net/mlx5e: Refactor doorbell function to allow avoiding a completion
Currently the doorbell function always asks for completion to be generated.

Refactor things such that all existing call sites are untouched and no
branching is added. This is done using inner function which can be invoked
directly in cases completion is not desired (as done in downstream patch).

No functional change here.

Issue: 2151469
Change-Id: I8db8c683c5417c4f8de1464243ac608492a3a283
Signed-off-by: Or Gerlitz <ogerlitz@nvidia.com>
Signed-off-by: Aurelien Aptel <aaptel@nvidia.com>